refactor(scripts): modularize generate-workflow-docs.js into 10 focused modules
SD-LEO-REFACTOR-WORKFLOW-DOCS-001

Extracts 1393 LOC monolith into domain-focused modules under scripts/workflow-docs-generator/:
- data-loader.js (75 LOC) - YAML loading and path constants
- stage-cards.js (61 LOC) - Individual stage diagram generation
- phase-diagrams.js (61 LOC) - Phase diagram generation
- sops.js (132 LOC) - SOP generation
- critiques.js (177 LOC) - Critique document generation
- prd-crosswalk.js (127 LOC) - PRD crosswalk generation
- backlog.js (300 LOC) - Backlog and issue generation
- research-packs.js (243 LOC) - Research brief generation
- validation-script.js (135 LOC) - Validation script generation
- readmes.js (150 LOC) - README generation
- index.js (64 LOC) - Main orchestration

Key improvements:
- All modules under 500 LOC (largest: 300 LOC)
- YAML caching via shared kernel pattern
- Re-export wrapper maintains backward compatibility
- No circular dependencies (flat module structure)
