+++ scripts/workflow-docs-generator/validation-script.js
@@ -1,0 +1,135 @@
+/**
+ * Validation Script Generator
+ * Generates the workflow validation script
+ *
+ * Extracted from generate-workflow-docs.js for modularity
+ * SD-LEO-REFACTOR-WORKFLOW-DOCS-001
+ */
+
+import fs from 'fs';
+import path from 'path';
+import { scriptsDir } from './data-loader.js';
+
+/**
+ * Generate validation script for workflow consistency checks
+ */
+export function generateValidationScript() {
+  const scriptPath = path.join(scriptsDir, 'validate-stages.js');
+  const scriptContent = `#!/usr/bin/env node
+/**
+ * Workflow Validation Script
+ * Validates stages.yaml for consistency and completeness
+ *
+ * Generated by workflow-docs-generator
+ */
+
+import fs from 'fs';
+import path from 'path';
+import yaml from 'js-yaml';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import { glob } from 'glob';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+// Load stages data
+const stagesPath = path.join(__dirname, '../docs/workflow/stages.yaml');
+const stagesData = yaml.load(fs.readFileSync(stagesPath, 'utf8'));
+const stages = stagesData.stages;
+
+let errors = [];
+let warnings = [];
+let missing = [];
+
+// Check for all 40 stages
+for (let i = 1; i <= 40; i++) {
+  const stage = stages.find(s => s.id === i);
+  if (!stage) {
+    errors.push(\`Missing stage \${i}\`);
+  }
+}
+
+// Validate dependencies
+stages.forEach(stage => {
+  stage.depends_on.forEach(dep => {
+    if (!stages.find(s => s.id === dep)) {
+      errors.push(\`Stage \${stage.id} depends on non-existent stage \${dep}\`);
+    }
+    if (dep >= stage.id) {
+      warnings.push(\`Stage \${stage.id} depends on later stage \${dep} (possible circular dependency)\`);
+    }
+  });
+});
+
+// Check for required files
+stages.forEach(stage => {
+  const padded = String(stage.id).padStart(2, '0');
+
+  // Check for SOP
+  const sopPattern = path.join(__dirname, \`../docs/workflow/sop/\${padded}-*.md\`);
+  const sopFiles = glob.sync(sopPattern);
+  if (sopFiles.length === 0) {
+    missing.push(\`SOP for stage \${stage.id}\`);
+  }
+
+  // Check for individual diagram
+  const diagramPath = path.join(__dirname, \`../docs/stages/individual/\${padded}.mmd\`);
+  if (!fs.existsSync(diagramPath)) {
+    missing.push(\`Diagram for stage \${stage.id}\`);
+  }
+
+  // Check for critique
+  const critiquePath = path.join(__dirname, \`../docs/workflow/critique/stage-\${padded}.md\`);
+  if (!fs.existsSync(critiquePath)) {
+    missing.push(\`Critique for stage \${stage.id}\`);
+  }
+});
+
+// Check for TBDs
+stages.forEach(stage => {
+  if (stage.inputs.includes('TBD') || stage.outputs.includes('TBD')) {
+    warnings.push(\`Stage \${stage.id} has TBD inputs/outputs\`);
+  }
+  if (stage.metrics.includes('TBD')) {
+    warnings.push(\`Stage \${stage.id} has TBD metrics\`);
+  }
+});
+
+// Report results
+console.log('\\n=== Workflow Validation Report ===\\n');
+
+if (errors.length > 0) {
+  console.log('ERRORS:');
+  errors.forEach(e => console.log(\`   - \${e}\`));
+} else {
+  console.log('No errors found');
+}
+
+if (warnings.length > 0) {
+  console.log('\\nWARNINGS:');
+  warnings.forEach(w => console.log(\`   - \${w}\`));
+} else {
+  console.log('\\nNo warnings');
+}
+
+if (missing.length > 0) {
+  console.log('\\nMISSING FILES:');
+  missing.forEach(m => console.log(\`   - \${m}\`));
+} else {
+  console.log('\\nAll required files present');
+}
+
+console.log(\`\\n=== Summary ===\`);
+console.log(\`Stages: \${stages.length}/40\`);
+console.log(\`Errors: \${errors.length}\`);
+console.log(\`Warnings: \${warnings.length}\`);
+console.log(\`Missing files: \${missing.length}\`);
+
+process.exit(errors.length > 0 ? 1 : 0);
+`;
+
+  fs.writeFileSync(scriptPath, scriptContent);
+  fs.chmodSync(scriptPath, '755');
+  console.log('Generated validation script');
+}
